requirements.txt
pandas
python-dotenv
open-interpreter
odfpy
Pyarrow